Revert "feat: 내 주변 기사님 찾기 지도 UI 추가"
