+++ messages/ko.json
@@ -212,10 +212,7 @@
       "uploadImageIconAlt": "이미지 업로드 아이콘",
       "dragDropArea": "이미지를 드래그하여 업로드하거나 클릭하여 선택하세요",
       "yearSuffix": "년",
-      "requiredField": "필수 입력 항목",
-      "businessLocationLabel": "사업장 위치",
-      "selectBusinessLocation": "사업장 위치 선택",
-      "confirmSelection": "선택 완료"
+      "requiredField": "필수 입력 항목"
    },
    "MoverSearch": {
       "findDriver": "기사님 찾기",
@@ -278,24 +275,6 @@
          "removedFromFavorites": "찜 목록에서 제거되었습니다."
       },
       "goToLogin": "로그인하러 가기",
-      "findOnMap": "지도에서 찾기",
-      "map": {
-         "mapSearchTitle": "내 주변 기사님 찾기",
-         "searchPlaceholder": "지역명을 입력하세요 (예: 강남구, 서울)",
-         "currentLocationBtn": "현재 위치로 이동",
-         "locationSearching": "위치 찾는 중...",
-         "mapLoadComplete": "지도 로드 완료",
-         "mapLoadFailed": "지도 로드 실패",
-         "mapLoadFailedDesc": "지도를 불러오는데 실패했습니다.",
-         "mapLoading": "지도를 불러오는 중...",
-         "refreshBtn": "새로고침",
-         "noMoversFound": "해당 지역에 기사님이 없습니다",
-         "searchOtherArea": "다른 지역을 검색해보세요",
-         "kakaoApiKeyMissing": "카카오맵 API 키가 설정되지 않았습니다.",
-         "moverDefaultName": "기사님",
-         "reviewCount": "리뷰",
-         "kmUnit": "km"
-      },
       "accessibility": {
          "driverCard": "{name} 기사님 카드",
          "loading": "로딩 중...",
