+++ package.json
@@ -25,7 +25,6 @@
       "jose": "^6.0.12",
       "lodash": "^4.17.21",
       "lottie-react": "^2.4.1",
-      "lucide-react": "^0.539.0",
       "next": "15.3.4",
       "next-intl": "^4.3.4",
       "nodemailer": "^7.0.5",
