+++ src/constants/mover.constants.ts
@@ -1,3 +1,5 @@
+// mover.constants.ts
+
 import { DropdownOption } from "@/lib/types/mover.types";
 
 export const AREA_OPTIONS: DropdownOption[] = [
@@ -35,101 +37,24 @@ export const SORT_OPTIONS: DropdownOption[] = [
    { label: "확정 많은순", value: "mostBooked" },
 ];
 
+
 // 지역 코드 → 실제 이름 매핑
 export const areaMapping: { [key: string]: string[] } = {
-   seoul: ["서울"],
-   incheon: ["인천"],
-   daejeon: ["대전"],
-   daegu: ["대구"],
-   gwangju: ["광주"],
-   busan: ["부산"],
-   ulsan: ["울산"],
-   sejong: ["세종"],
-   gyeonggi: ["경기"],
-   gangwon: ["강원"],
-   chungbuk: ["충북"],
-   chungnam: ["충남"],
-   jeonbuk: ["전북"],
-   jeonnam: ["전남"],
-   gyeongbuk: ["경북"],
-   gyeongnam: ["경남"],
-   jeju: ["제주"],
-};
-
-// 기본 좌표 상수
-export const DEFAULT_COORDINATES: Record<string, { lat: number; lng: number }> =
-   {
-      서울: { lat: 37.5665, lng: 126.978 },
-      인천: { lat: 37.4563, lng: 126.7052 },
-      대전: { lat: 36.3504, lng: 127.3845 },
-      대구: { lat: 35.8714, lng: 128.6014 },
-      광주: { lat: 35.1595, lng: 126.8526 },
-      부산: { lat: 35.1796, lng: 129.0756 },
-      울산: { lat: 35.5384, lng: 129.3114 },
-      세종: { lat: 36.48, lng: 127.289 },
-      경기: { lat: 37.4138, lng: 127.5183 },
-      강원: { lat: 37.8228, lng: 128.1555 },
-      충북: { lat: 36.8, lng: 127.7 },
-      충남: { lat: 36.5, lng: 126.8 },
-      전북: { lat: 35.7175, lng: 127.153 },
-      전남: { lat: 34.8679, lng: 126.991 },
-      경북: { lat: 36.4919, lng: 128.888 },
-      경남: { lat: 35.4606, lng: 128.2132 },
-      제주: { lat: 33.4996, lng: 126.5312 },
-   };
-
-// 지역명 매핑 상수
-export const REGION_MAP: Record<string, string> = {
-   서울: "서울",
-   서울시: "서울",
-   서울특별시: "서울",
-   부산: "부산",
-   부산시: "부산",
-   부산광역시: "부산",
-   대구: "대구",
-   대구시: "대구",
-   대구광역시: "대구",
-   인천: "인천",
-   인천시: "인천",
-   인천광역시: "인천",
-   광주: "광주",
-   광주시: "광주",
-   광주광역시: "광주",
-   대전: "대전",
-   대전시: "대전",
-   대전광역시: "대전",
-   울산: "울산",
-   울산시: "울산",
-   울산광역시: "울산",
-   세종: "세종",
-   세종시: "세종",
-   세종특별자치시: "세종",
-   경기: "경기",
-   경기도: "경기",
-   강원: "강원",
-   강원도: "강원",
-   충북: "충북",
-   충청북도: "충북",
-   충남: "충남",
-   충청남도: "충남",
-   전북: "전북",
-   전라북도: "전북",
-   전북특별자치도: "전북",
-   전남: "전남",
-   전라남도: "전남",
-   경북: "경북",
-   경상북도: "경북",
-   경남: "경남",
-   경상남도: "경남",
-   제주: "제주",
-   제주도: "제주",
-   제주특별자치도: "제주",
-   강남구: "서울",
-   서초구: "서울",
-   송파구: "서울",
-   강서구: "서울",
-   마포구: "서울",
-   종로구: "서울",
-   중구: "서울",
-   영등포구: "서울",
+  seoul: ["서울"],
+  incheon: ["인천"],
+  daejeon: ["대전"],
+  daegu: ["대구"],
+  gwangju: ["광주"],
+  busan: ["부산"],
+  ulsan: ["울산"],
+  sejong: ["세종"],
+  gyeonggi: ["경기"],
+  gangwon: ["강원"],
+  chungbuk: ["충북"],
+  chungnam: ["충남"],
+  jeonbuk: ["전북"],
+  jeonnam: ["전남"],
+  gyeongbuk: ["경북"],
+  gyeongnam: ["경남"],
+  jeju: ["제주"],
 };
