+++ src/constants/profile.constants.ts
@@ -1,4 +1,3 @@
-//profile.constants.ts
 export const regions = [
    "서울",
    "경기",
