+++ src/lib/api/auth/requests/updateMoverProfile.ts
@@ -12,5 +12,6 @@ export default async function updateMoverProfile(
    });
 
    tokenSettings.set(res.accessToken);
+
    return { ...res };
 }
