+++ src/lib/types/client.types.ts
@@ -1,4 +1,3 @@
-//frontend client.types.ts
 import { FieldValues, Path, UseFormRegister } from "react-hook-form";
 
 // 서비스 종류
